+++ src/IOrdenador.java
@@ -1,6 +1,6 @@
 public interface IOrdenador<T>{
     public T[] ordenar(T[] dados);
-    public int getComparacoes();
-    public int getMovimentacoes();
+    public long getComparacoes();
+    public long getMovimentacoes();
     public double getTempoOrdenacao();
 }
